Merge Blog Backend

+++ src/main/java/com/guico/moviemanagesystembackend/utils/Result.java
@@ -13,7 +13,7 @@ public class Result {
     private Boolean success;
     private String errorMsg;
     private Object data;
-    private Long total;
+    private Integer total;
 
     public static Result ok() {
         return new Result(true, null, null, null);
@@ -23,7 +23,7 @@ public class Result {
         return new Result(true, null, data, null);
     }
 
-    public static Result ok(List<?> data, Long total) {
+    public static Result ok(List<?> data, Integer total) {
         return new Result(true, null, data, total);
     }
